refactor: align SelectedSearchPlace with SearchResultPlace for unified place selection logic
- Updated SelectedSearchPlace to manage and receive Place object via props and internal state.
- Enabled future reuse of common APIs and UI logic between search result and selected display.
- Deferred search and selection component refactors to next steps.
