+++ prop-drilling-shop-maximillian/src/App.tsx
@@ -2,21 +2,13 @@ import { useState } from "react";
 import "./App.css";
 import SearchPlace from "./components/SearchPlace";
 import SelectedPlace from "./components/SelectedPlace";
-import SearchFilter from "./components/SearchFilter";
-import type { AvailablePlace } from "./types/AvailablePlace";
+import SelectedSearchPlace from "./components/SelectedSearchPlace";
 
 function App() {
   const [searchTitle, setSearchTitle] = useState("");
-  const [selectedPlace, setSelectedPlace] = useState<AvailablePlace | null>(
-    null
-  );
 
   const handleSearch = (title: string) => {
     setSearchTitle(title);
-  };
-
-  const handleOnSelectPlace = (place: AvailablePlace) => {
-    setSelectedPlace(place);
   };
 
   return (
@@ -26,16 +18,7 @@ function App() {
       {searchTitle && (
         <p className="search-summary">You searched: {searchTitle}</p>
       )}
-      {searchTitle && (
-        <SearchFilter
-          searchTerm={searchTitle}
-          onSelectPlace={handleOnSelectPlace}
-        />
-      )}
-
-      {selectedPlace && (
-        <p>Selected in the Search Filter: {selectedPlace?.title}</p>
-      )}
+      <SelectedSearchPlace searchTitle={searchTitle} />
     </>
   );
 }
