+++ prop-drilling-shop-maximillian/src/components/SearchResultPlace.tsx
@@ -4,13 +4,15 @@ import type { SearchResultPlaceProps } from "../types/AvailablePlace";
 const SearchResultPlace: React.FC<SearchResultPlaceProps> = ({
   place,
   onSelectPlace,
+  isSelected,
 }) => {
   const handleOnSelectPlace = () => {
     onSelectPlace(place);
   };
   return (
     <div>
-      <li className="place-item">
+      //Use conditional styling inside Place.tsx
+      <li className={`place-item ${isSelected ? "selected" : ""}`}>
         <button onClick={handleOnSelectPlace}>
           <img src={place.image.src} alt={place.image.alt} />
           <h3>{place.title}</h3>
